This adds upgrade functionality so that existing django apps can switch to SRP.
If a user exists in the auth table but not the srp table, the server sends back the algorithm and salt needed to hash the password. The hashed 
password is used to authenticate the user.

After the server authenticates the user and the user verifies the identity of the server, the user sends the password in plaintext. The server 
uses the plaintext password to calculate the verifier and stores. Finally, the client reinitiates the login process.